almost compiling on 64bit...
git-svn-id: file:///afs/psi.ch/project/sls_det_software/svn/slsDetectorSoftware@297 951219d9-93cf-4727-9268-0efd64621fa3

+++ slsDetectorSoftware/multiSlsDetector/multiSlsDetector.cpp
@@ -2194,7 +2194,7 @@ dacs_t multiSlsDetector::getADC(dacIndex idac, int imod) {
   return ret1;
 }
 
-int multiSlsDetector::setChannel(long long reg, int ichan, int ichip, int imod) {
+int multiSlsDetector::setChannel(int64_t reg, int ichan, int ichip, int imod) {
   int ret, ret1=-100;
   int id=-1, im=-1;
   int dmi=0, dma=thisMultiDetector->numberOfDetectors;
